requirements.txt
Flask==2.3.3 
mysql-connector-python==8.1.0 
pandas==1.5.3 
numpy==1.24.3 
openpyxl==3.1.2 
Werkzeug==2.3.7 
